+++ apps/opplaringspenger-soknad/src/app/søknad/steps/kurs/kursStepUtils.ts
@@ -10,6 +10,7 @@ import {
     getDatesInDateRange,
     getDatesInDateRanges,
     isDateInDateRanges,
+    isDateWeekDay,
     sortDateRange,
 } from '@navikt/sif-common-utils';
 import { getDateRangeValidator, getListValidator } from '@navikt/sif-validation';
@@ -251,6 +252,10 @@ export const getDatoerUtenforSøknadsperioder = (datoer: Date[], søknadsperiode
     return datoer.filter((reisedag) => !isDateInDateRanges(reisedag, søknadsperioder));
 };
 
+export const getDatoerSomErHelg = (datoer: Date[]): Date[] => {
+    return datoer.filter((reisedag) => isDateWeekDay(reisedag) === false);
+};
+
 export const erAlleReisedagerInnenforSøknadsperioder = (
     reisedager: Enkeltdato[],
     søknadsperioder: DateRange[],
@@ -263,6 +268,17 @@ export const getReisedagerValidator = (kursperioder: DateRange[]) => {
         const error = getListValidator({ required: true })(reisedager);
         if (error) {
             return error;
+        }
+        /** Kontroller om datoer er på en helgedag */
+        const reisedagerPåHelg = getDatoerSomErHelg(reisedager.map((d) => d.dato));
+        if (reisedagerPåHelg.length > 0) {
+            return {
+                key: 'reisedagPåHelg',
+                values: {
+                    antallDager: reisedagerPåHelg.length,
+                    dager: reisedagerPåHelg.map((d) => dateFormatter.dayCompactDate(d)).join(', '),
+                },
+            };
         }
         /** Kontroller om datoer er innenfor søknadsperioder */
         const reisedagerUtenforSøknadsperioder = getDatoerUtenforSøknadsperioder(
@@ -330,4 +346,14 @@ export const getUtenlandsoppholdValidator = (kursperioder: DateRange[]) => {
         }
         return undefined;
     };
+};
+
+export const startOgSluttErSammeHelg = (start?: Date, slutt?: Date): boolean => {
+    if (!start || !slutt) return false;
+
+    const startErHelgedag = !isDateWeekDay(start);
+    const sluttErHelgedag = !isDateWeekDay(slutt);
+    const dagerMellom = Math.abs(dayjs(slutt).diff(dayjs(start), 'day'));
+
+    return startErHelgedag && sluttErHelgedag && (dagerMellom === 0 || dagerMellom === 1);
 };