+++ apps/ungdomsytelse-deltaker/src/apps/innsyn/i18n/messages/nb.ts
@@ -1,4 +1,3 @@
-/* eslint-disable max-len */
 export const innsynMessages_nb = {
     'innsyn.sidetittel': 'Din ungdomsprogramytelse',
     Ja: 'Ja',
@@ -7,8 +6,8 @@ export const innsynMessages_nb = {
 
     'ytelse.SYKEPENGER': 'Sykepenger',
     'ytelse.OMSORGSPENGER': 'Omsorgspenger',
-    'ytelse.PLEIEPENGER_SYKT_BARN': 'Pleiepenger for sykt barn',
-    'ytelse.PLEIEPENGER_LIVETS_SLUTTFASE': 'Pleiepenger i livets sluttfase',
+    'ytelse.PLEIEPENGER_SYKT_BARN': 'Pleiepenger',
+    'ytelse.PLEIEPENGER_LIVETS_SLUTTFASE': 'Pleiepenger',
     'ytelse.OPPLAERINGSPENGER': 'Opplæringspenger',
 
     'forsideLenkeButton.tekst': 'Tilbake til oversikten',
@@ -52,7 +51,7 @@ export const innsynMessages_nb = {
     'deltakelseIkkeStartetInfo.tekst2':
         'På denne siden får du oppgaver mens du er i ungdomsprogrammet. Du trenger ikke gjøre noe med oppgavene før du får et varsel på SMS om det.',
     'deltakelsePågåendeInfo.tekst':
-        'Husk å melde fra innen 6. hver måned hvis du jobbet og fikk lønn måneden før. Du får et varsel på SMS når du skal melde fra.',
+        'Husk å melde fra senest 7. hver måned hvis du jobbet og fikk lønn måneden før. Du får et varsel på SMS når du skal melde fra.',
 
     'oppgaveOgTilbakemelding.header': 'Beskjed og ditt svar',
     'oppgaveOgTilbakemelding.beskjedFraNav': 'Beskjed fra Nav',
@@ -61,6 +60,9 @@ export const innsynMessages_nb = {
     'oppgavebekreftelse.ubesvart.tittel': 'Hei {deltakerNavn}',
     'oppgavebekreftelse.kvittering.tittel': 'Svaret ditt er sendt inn',
     'oppgavebekreftelse.besvart.svarMangler': 'Informasjon om hva du svarte er ikke tilgjengelig enda.',
+
+    'oppgavebekreftelse.oppgavetekst.ariaLabel': 'Oppgaveinformasjon',
+    'oppgavebekreftelse.uttalelseform.ariaLabel': 'Svarskjema',
 
     /** OppgaveMessages */
     'oppgavestatus.LØST': 'Løst',
@@ -74,6 +76,8 @@ export const innsynMessages_nb = {
     'oppgavetype.BEKREFT_ENDRET_STARTDATO.info':
         'Veilederen din har endret datoen for når du startet i ungdomsprogrammet.',
     'oppgavetype.BEKREFT_ENDRET_STARTDATO.harTilbakemeldingSpørsmål': 'Har du en tilbakemelding på startdatoen?',
+    'oppgavetype.BEKREFT_ENDRET_STARTDATO.harIkkeUttalelseLabel': 'Nei',
+    'oppgavetype.BEKREFT_ENDRET_STARTDATO.harUttalelseLabel': 'Ja',
     'oppgavetype.BEKREFT_ENDRET_STARTDATO.tilbakemeldingFritekstLabel': 'Tilbakemelding',
     'oppgavetype.BEKREFT_ENDRET_STARTDATO.kvitteringTekst': 'Du vil om kort tid motta et oppdatert vedtak.',
 
@@ -81,6 +85,8 @@ export const innsynMessages_nb = {
     'oppgavetype.BEKREFT_MELDT_UT.oppgavetittel': 'Tilbakemelding på sluttdato',
     'oppgavetype.BEKREFT_MELDT_UT.info': 'Veilederen din har satt en dato for når du slutter i ungdomsprogrammet.',
     'oppgavetype.BEKREFT_MELDT_UT.harTilbakemeldingSpørsmål': 'Har du en tilbakemelding på sluttdatoen?',
+    'oppgavetype.BEKREFT_MELDT_UT.harIkkeUttalelseLabel': 'Nei',
+    'oppgavetype.BEKREFT_MELDT_UT.harUttalelseLabel': 'Ja',
     'oppgavetype.BEKREFT_MELDT_UT.tilbakemeldingFritekstLabel': 'Tilbakemelding',
     'oppgavetype.BEKREFT_MELDT_UT.kvitteringTekst': 'Du vil om kort tid motta et oppdatert vedtak.',
 
@@ -89,14 +95,19 @@ export const innsynMessages_nb = {
     'oppgavetype.BEKREFT_ENDRET_SLUTTDATO.info':
         'Veilederen din har endret datoen for når du slutter i ungdomsprogrammet.',
     'oppgavetype.BEKREFT_ENDRET_SLUTTDATO.harTilbakemeldingSpørsmål': 'Har du en tilbakemelding på sluttdatoen?',
+    'oppgavetype.BEKREFT_ENDRET_SLUTTDATO.harIkkeUttalelseLabel': 'Nei',
+    'oppgavetype.BEKREFT_ENDRET_SLUTTDATO.harUttalelseLabel': 'Ja',
     'oppgavetype.BEKREFT_ENDRET_SLUTTDATO.tilbakemeldingFritekstLabel': 'Tilbakemelding',
     'oppgavetype.BEKREFT_ENDRET_SLUTTDATO.kvitteringTekst': 'Du vil om kort tid motta et oppdatert vedtak.',
 
     'oppgavetype.BEKREFT_AVVIK_REGISTERINNTEKT.paneltittel': 'Sjekk lønn i {månedOgÅr}',
     'oppgavetype.BEKREFT_AVVIK_REGISTERINNTEKT.oppgavetittel': 'Tilbakemelding på lønn i {månedOgÅr}',
     'oppgavetype.BEKREFT_AVVIK_REGISTERINNTEKT.info':
-        'Sjekk at opplysningene våre om lønnen du fikk i {måned} er riktige. Send oss et svar når du har sjekket.',
-    'oppgavetype.BEKREFT_AVVIK_REGISTERINNTEKT.harTilbakemeldingSpørsmål': 'Har du en tilbakemelding?',
+        'Sjekk at opplysningene våre om lønnen du fikk fra arbeidsgiver i {måned} er riktige. Send oss et svar når du har sjekket.',
+    'oppgavetype.BEKREFT_AVVIK_REGISTERINNTEKT.harTilbakemeldingSpørsmål': 'Stemmer lønnen vi har fått oppgitt?',
+    'oppgavetype.BEKREFT_AVVIK_REGISTERINNTEKT.harIkkeUttalelseLabel': 'Ja, lønnen stemmer',
+    'oppgavetype.BEKREFT_AVVIK_REGISTERINNTEKT.harUttalelseLabel':
+        'Nei, lønnen stemmer ikke. Jeg har en tilbakemelding.',
     'oppgavetype.BEKREFT_AVVIK_REGISTERINNTEKT.tilbakemeldingFritekstLabel': 'Tilbakemelding',
     'oppgavetype.BEKREFT_AVVIK_REGISTERINNTEKT.kvitteringTekst':
         'Vi bruker tilbakemeldingen din når vi vurderer hvor mye penger du skal få i {utbetalingsmåned}. Du får et vedtaksbrev om dette.',
@@ -117,9 +128,17 @@ export const innsynMessages_nb = {
     'inntektForm.utbetaltLønnLegend': 'Fikk du utbetalt lønn i {måned}?',
     'inntektForm.lønnLabel': 'Hvor mye fikk du i lønn før skatt?',
     'inntektForm.lønnDescription': 'Se på lønnsslippen din hva lønnen din var før det ble trukket skatt av den.',
-    'inntektForm.hvordanFinnerDuUtLønn': 'Hvordan finner du ut hva lønnen din var før skatt?',
-    'inntektForm.hvordanFinnerDuUtLønnBeskrivelse':
-        'Når du har en jobb der du får utbetalt lønn, får du alltid en lønnsslipp fra arbeidsgiveren din. På lønnsslippen står det blant annet hva lønnen din er før det blir trukket skatt av den, og det er det tallet du skal skrive inn her. Spør arbeidsgiveren din hvis du er usikker på hvor du finner lønnsslippen.',
+    'inntektForm.hvordanFinnerDuUtLønn.tittel': 'Hvordan finner du ut hva lønnen din var før skatt?',
+    'inntektForm.hvordanFinnerDuUtLønnBeskrivelse.tekst.1':
+        'Når du har en jobb der du får utbetalt lønn, får du alltid en lønnsslipp fra arbeidsgiveren din. På lønnsslippen står det blant annet hva lønnen din er før det blir trukket skatt av den, og det er det tallet du skal skrive inn her.',
+    'inntektForm.hvordanFinnerDuUtLønnBeskrivelse.tekst.2':
+        'Spør arbeidsgiveren din hvis du er usikker på hvor du finner lønnsslippen.',
+    'inntektForm.feilLønn.tittel': 'Hva skjer hvis jeg melder inn feil lønn?',
+    'inntektForm.feilLønn.tekst.1':
+        'Vi sjekker alltid hva arbeidsgiveren din har registrert at du fikk i lønn. Arbeidsgivere registrerer lønn i et system som heter A-ordningen og som blant annet Nav og Skatteetaten har tilgang til.',
+    'inntektForm.feilLønn.tekst.2':
+        'Hvis det er forskjell på lønnen din i A-ordningen og lønnen du har sendt inn til oss, får du beskjed om det.',
+
     'inntektForm.hentUtBeløpFeil':
         'Du har oppgitt at du fikk utbetalt lønn, men vi klarer ikke hente ut beløpet. Vennligst kontroller feltet for lønn og prøv på nytt.',
 
@@ -131,11 +150,10 @@ export const innsynMessages_nb = {
     'inntektForm.validation.lønn.numberHasDecimals': 'Du må oppgi lønn uten desimaler.',
 
     /** UtalelseForm */
-    'uttalelseForm.ariaLabel': 'Skjema for tilbakemelding',
     'uttalelseForm.submitButtonLabel': 'Send inn svaret ditt',
     'uttalelseForm.cancelButtonLabel': 'Avbryt',
     'uttalelseForm.defaultDescription':
-        'Du må ikke oppgi sensitive informasjon (særlige kategorier av personopplysninger) om deg selv eller andre, for eksempel helseopplysninger.',
+        'Du må ikke oppgi sensitive informasjon (personopplysninger) om deg selv eller andre, for eksempel helseopplysninger.',
 
     'uttalelseForm.validation.harTilbakemelding.yesOrNoIsUnanswered': 'Du må svare på om du har en tilbakemelding.',
     'uttalelseForm.validation.tilbakemelding.stringHasNoValue': 'Du må fylle ut tilbakemeldingsfeltet.',