+++ apps/ungdomsytelse-deltaker/src/apps/innsyn/modules/oppgaver/avvik-registerinntekt/i18n/nb.ts
@@ -2,28 +2,35 @@ export const avvikRegisterinntektMessages_nb = {
     'avvikRegisterinntekt.inntekskilde.arbeidsgiverYtelse': 'Arbeidsgiver/Nav-ytelse',
     'avvikRegisterinntekt.inntekskilde.arbeidsgiver': 'Arbeidsgiver',
     'avvikRegisterinntekt.inntekskilde.navytelse': 'Nav-ytelse',
-    'avvikRegisterinntektOppsummering.intro': 'Vi har fått disse opplysningene om lønnen din i {rapporteringsmåned}:',
+    'avvikRegisterinntektOppsummering.intro':
+        'Vi har fått disse opplysningene fra arbeidsgiver om lønnen din i {rapporteringsmåned}:',
 
     'inntektTabell.totalt': 'Totalt',
     'inntektTabell.lønn': 'Lønn (før skatt)',
 
     'avvikRegisterinntektOppgavetekst.navYtelse':
-        'Vi har fått disse opplysningene om ytelse fra Nav i {rapporteringsmåned}:',
+        'Vi har fått disse opplysningene om ytelse fra Nav for {rapporteringsmåned}:',
     'avvikRegisterinntektOppgavetekst.generell':
-        'Vi har fått disse opplysningene om lønnen din i {rapporteringsmåned}:',
+        'Vi har fått disse opplysningene fra arbeidsgiver om lønnen din for {rapporteringsmåned}:',
     'avvikRegisterinntektOppgavetekst.ingenOpplysninger':
         'Du har gitt oss beskjed om at du hadde lønn i {rapporteringsmåned}, men vi har ikke fått inn opplysninger fra arbeidsgiver om at du hadde lønn i {rapporteringsmåned}.',
-    'avvikRegisterinntektOppgavetekst.1':
-        'Før vi vurderer hvor mye penger du får utbetalt i {utbetalingsmåned}, kan du komme med en tilbakemelding på lønnen for {rapporteringsmåned}.',
-    'avvikRegisterinntektOppgavetekst.2': 'Hvis du ikke har en tilbakemelding, krysser du av på “Nei”.',
-    'avvikRegisterinntektOppgavetekst.3.harInntekt':
-        'Hvis du ser at lønnen er feil, sjekker du den med arbeidsgiveren din først. Hvis du fortsatt mener at den er feil, krysser du av på “Ja” og sender en tilbakemelding til oss om det.',
-    'avvikRegisterinntektOppgavetekst.3.harIkkeInntekt':
-        'Hvis du likevel mener at du hadde lønn i {rapporteringsmåned}, krysser du av på “Ja” og sender en tilbakemelding til oss om det.',
-    'avvikRegisterinntektOppgavetekst.4': 'Jo fortere du svarer, jo fortere får du pengene utbetalt.',
-    'avvikRegisterinntektOppgavetekst.5': 'Fristen for å svare er {formatertFrist}.',
-    'avvikRegisterinntektOppgavetekst.6':
-        'Hvis vi ikke hører fra deg innen svarfristen har gått ut, bruker vi lønnen som arbeidsgiver har oppgitt.',
+    'avvikRegisterinntektOppgavetekst.ingenInntekt':
+        'Vi bruker opplysningene fra arbeidsgiver når vi vurdere hvor mye du får utbetalt. Når vi ikke har mottatt noe fra arbeidsgiver, vil vi basere oss på at du ikke fikk utbetalt lønn i {rapporteringsmåned}.',
+    'avvikRegisterinntektOppgavetekst.1.harInntekt':
+        'Vi bruker denne lønnen fra arbeidsgiver til å vurdere hvor mye du får utbetalt.',
+    'avvikRegisterinntektOppgavetekst.1.harInntekt.kunYtelse':
+        'Vi bruker denne lønnen til å vurdere hvor mye du får utbetalt.',
+    'avvikRegisterinntektOppgavetekst.2': 'Vi trenger et svar fra deg på om lønnen stemmer eller ikke.',
+    'avvikRegisterinntektOppgavetekst.3': 'Hvis lønnen stemmer, krysser du av for <strong>Ja, lønnen stemmer</strong>.',
+    'avvikRegisterinntektOppgavetekst.4':
+        'Hvis du mener at lønnen er feil, krysser du av på <strong>Nei, lønnen stemmer ikke</strong> og sender en tilbakemelding til oss om det.',
+
+    'avvikRegisterinntektOppgavetekst.5': 'Jo fortere du svarer, jo fortere får du pengene utbetalt.',
+    'avvikRegisterinntektOppgavetekst.6': 'Fristen for å svare er {formatertFrist}.',
+    'avvikRegisterinntektOppgavetekst.7':
+        'Hvis vi ikke hører fra deg innen svarfristen, bruker vi lønnen som arbeidsgiver har oppgitt.',
+    'avvikRegisterinntektOppgavetekst.7.kunYtelse':
+        'Hvis vi ikke hører fra deg innen svarfristen, bruker vi lønnen vi har fått har oppgitt.',
 
     'regelverkOgInnsyn.readMore.tittel': 'Regelverk og innsyn',
     'regelverkOgInnsyn.readMore.tekst.1': 'Se regelverket for ungdomsprogramytelsen:',