+++ apps/ungdomsytelse-deltaker/src/apps/innsyn/utils/deltakelseUtils.ts
@@ -1,5 +1,7 @@
 import { getDateToday } from '@navikt/sif-common-utils';
+import { OppgaveStatus, Oppgavetype } from '@navikt/ung-deltakelse-opplyser-api-deltaker';
 import { DeltakelsePeriode } from '@shared/types/DeltakelsePeriode';
+import { Oppgave } from '@shared/types/Oppgave';
 import dayjs from 'dayjs';
 
 /**
@@ -26,4 +28,13 @@ export const erDeltakelseStartet = (deltakelsePeriode: DeltakelsePeriode): boole
         return false;
     }
     return true;
+};
+
+export const harRapportertInntekt = (oppgaver: Oppgave[]): boolean => {
+    return oppgaver.some(
+        (o) =>
+            o.oppgavetype === Oppgavetype.RAPPORTER_INNTEKT &&
+            o.status !== OppgaveStatus.ULØST &&
+            o.oppgavetypeData?.rapportertInntekt !== undefined,
+    );
 };