+++ apps/ungdomsytelse-deltaker/src/apps/innsyn/ux-signals/UXRapportertInntekt.tsx
@@ -1,0 +1,9 @@
+import { useUxSignalsLoader, UxSignalsPanel } from '@navikt/sif-common-core-ds';
+
+const UXRapportertInntekt = () => {
+    useUxSignalsLoader(true);
+
+    return <UxSignalsPanel panelId="1bmhfberhs" />;
+};
+
+export default UXRapportertInntekt;