+++ apps/ungdomsytelse-deltaker/src/utils/sentryUtils.ts
@@ -7,7 +7,7 @@ enum SentryEnvironment {
 }
 
 const resolveEnvironment = (): SentryEnvironment | undefined => {
-    const hostname = window.location.hostname || undefined;
+    const hostname = globalThis.location.hostname || undefined;
     if (hostname) {
         if (hostname.includes('localhost')) {
             return SentryEnvironment.localhost;
@@ -29,6 +29,6 @@ export const initSentry = () => {
         dsn: 'https://01c0cdacd803d88882c2eab4c345c610@sentry.gc.nav.no/179',
         integrations: [Sentry.browserTracingIntegration()],
         environment,
-        tracesSampleRate: 0.0, //  Skru av default tracing
+        tracesSampleRate: 0, //  Skru av default tracing
     });
 };