+++ apps/ungdomsytelse-deltaker/vite.dev.config.ts
@@ -33,7 +33,8 @@ export default defineConfig({
     },
     define: {
         __IS_GITHUB_PAGES__: false,
-        __INJECT_DECORATOR_CLIENT_SIDE__: true,
+        __IS_VEILEDER_DEMO__: true,
+        __INJECT_DECORATOR_CLIENT_SIDE__: false,
     },
     server: {
         port: 8080,