+++ packages/k9-brukerdialog-prosessering-api/CHANGELOG.md
@@ -1,5 +1,51 @@
 # @navikt/k9-brukerdialog-prosessering-api
 
+## 0.2.0
+
+### Minor Changes
+
+- Codegen
+
+### Patch Changes
+
+- Pakkeoppdateringer - minor/patch
+
+## 0.1.24
+
+### Patch Changes
+
+- Pakkeoppdateringer - minor/patch
+
+## 0.1.23
+
+### Patch Changes
+
+- Pakkeoppdateringer - minor/patch
+
+## 0.1.22
+
+### Patch Changes
+
+- 52d9b9e: Pakkeoppdateringer - minor/patch
+
+## 0.1.21
+
+### Patch Changes
+
+- Pakkeoppdateringer - minor/patch
+
+## 0.1.20
+
+### Patch Changes
+
+- Pakkeoppdateringer - minor/patch
+
+## 0.1.19
+
+### Patch Changes
+
+- Pakkeoppdataringer - minor/patch
+
 ## 0.1.18
 
 ### Patch Changes