+++ packages/sif-app-register/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @navikt/sif-app-register
 
+## 0.1.10
+
+### Patch Changes
+
+- Pakkeoppdateringer - minor/patch
+
 ## 0.1.9
 
 ### Patch Changes