+++ packages/sif-common-utils/CHANGELOG.md
@@ -1,5 +1,53 @@
 # @navikt/sif-common-utils
 
+## 3.49.39
+
+### Patch Changes
+
+- Pakkeoppdateringer - minor/patch
+
+## 3.49.38
+
+### Patch Changes
+
+- Pakkeoppdateringer - minor/patch
+
+## 3.49.37
+
+### Patch Changes
+
+- Pakkeoppdateringer - minor/patch
+
+## 3.49.36
+
+### Patch Changes
+
+- 52d9b9e: Pakkeoppdateringer - minor/patch
+
+## 3.49.35
+
+### Patch Changes
+
+- Pakkeoppdateringer - minor/patch
+
+## 3.49.34
+
+### Patch Changes
+
+- Pakkeoppdateringer - minor/patch
+
+## 3.49.33
+
+### Patch Changes
+
+- Pakkeoppdateringer - minor/patch. Fjerne ubrukte pakker.
+
+## 3.49.32
+
+### Patch Changes
+
+- Pakkeoppdataringer - minor/patch
+
 ## 3.49.31
 
 ### Patch Changes