+++ packages/ung-common/CHANGELOG.md
@@ -1,5 +1,86 @@
 # @navikt/ung-common
 
+## 0.0.59
+
+### Patch Changes
+
+- Updated dependencies
+- Updated dependencies
+    - @navikt/ung-deltakelse-opplyser-api-deltaker@0.1.33
+    - @navikt/ung-deltakelse-opplyser-api-veileder@0.1.26
+    - @navikt/k9-brukerdialog-prosessering-api@0.2.0
+
+## 0.0.58
+
+### Patch Changes
+
+- @navikt/ung-deltakelse-opplyser-api-deltaker@0.1.32
+
+## 0.0.57
+
+### Patch Changes
+
+- Pakkeoppdateringer - minor/patch
+- Updated dependencies
+    - @navikt/ung-deltakelse-opplyser-api-deltaker@0.1.31
+    - @navikt/ung-deltakelse-opplyser-api-veileder@0.1.25
+    - @navikt/k9-brukerdialog-prosessering-api@0.1.24
+
+## 0.0.56
+
+### Patch Changes
+
+- Updated dependencies
+    - @navikt/ung-deltakelse-opplyser-api-deltaker@0.1.30
+    - @navikt/ung-deltakelse-opplyser-api-veileder@0.1.24
+    - @navikt/k9-brukerdialog-prosessering-api@0.1.23
+
+## 0.0.55
+
+### Patch Changes
+
+- 52d9b9e: Pakkeoppdateringer - minor/patch
+- Updated dependencies [52d9b9e]
+    - @navikt/ung-deltakelse-opplyser-api-deltaker@0.1.29
+    - @navikt/ung-deltakelse-opplyser-api-veileder@0.1.23
+    - @navikt/k9-brukerdialog-prosessering-api@0.1.22
+
+## 0.0.54
+
+### Patch Changes
+
+- Updated dependencies
+    - @navikt/ung-deltakelse-opplyser-api-deltaker@0.1.28
+    - @navikt/ung-deltakelse-opplyser-api-veileder@0.1.22
+    - @navikt/k9-brukerdialog-prosessering-api@0.1.21
+
+## 0.0.53
+
+### Patch Changes
+
+- Pakkeoppdateringer - minor/patch
+- Updated dependencies
+    - @navikt/ung-deltakelse-opplyser-api-deltaker@0.1.27
+    - @navikt/ung-deltakelse-opplyser-api-veileder@0.1.21
+    - @navikt/k9-brukerdialog-prosessering-api@0.1.20
+
+## 0.0.52
+
+### Patch Changes
+
+- @navikt/k9-brukerdialog-prosessering-api@0.1.19
+- @navikt/ung-deltakelse-opplyser-api-deltaker@0.1.26
+- @navikt/ung-deltakelse-opplyser-api-veileder@0.1.20
+
+## 0.0.51
+
+### Patch Changes
+
+- Updated dependencies
+    - @navikt/ung-deltakelse-opplyser-api-deltaker@0.1.25
+    - @navikt/ung-deltakelse-opplyser-api-veileder@0.1.20
+    - @navikt/k9-brukerdialog-prosessering-api@0.1.19
+
 ## 0.0.50
 
 ### Patch Changes