+++ packages/ung-deltakelse-opplyser-api-veileder/CHANGELOG.md
@@ -1,5 +1,47 @@
 # @navikt/ung-deltakelse-opplyser-api
 
+## 0.1.26
+
+### Patch Changes
+
+- Pakkeoppdateringer - minor/patch
+
+## 0.1.25
+
+### Patch Changes
+
+- Pakkeoppdateringer - minor/patch
+
+## 0.1.24
+
+### Patch Changes
+
+- Pakkeoppdateringer - minor/patch
+
+## 0.1.23
+
+### Patch Changes
+
+- 52d9b9e: Pakkeoppdateringer - minor/patch
+
+## 0.1.22
+
+### Patch Changes
+
+- Pakkeoppdateringer - minor/patch
+
+## 0.1.21
+
+### Patch Changes
+
+- Pakkeoppdateringer - minor/patch
+
+## 0.1.20
+
+### Patch Changes
+
+- Pakkeoppdataringer - minor/patch
+
 ## 0.1.19
 
 ### Patch Changes